Started working on category screen composable. Moved previous composables to separate files.
