+++ app/build.gradle
@@ -73,6 +73,9 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-compose:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycle_version"
 
+    //Image loading library
+    implementation("io.coil-kt:coil-compose:2.2.0")
+
     //more icons
     implementation "androidx.compose.material:material-icons-extended:1.3.1"
