+++ app/src/main/java/com/example/shopapp/composable/ErrorTextFieldItem.kt
@@ -1,0 +1,37 @@
+package com.example.shopapp.composable
+
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.sp
+import com.example.shopapp.ui.theme.ShopAppTheme
+
+@Composable
+fun ErrorTextFieldItem(
+    errorMessage: String,
+    testTag: String
+) {
+    Text(
+        text = errorMessage,
+        fontSize = 13.sp,
+        color = MaterialTheme.colors.error,
+        textAlign = TextAlign.Start,
+        modifier = Modifier
+            .testTag(testTag)
+    )
+}
+
+@Preview
+@Composable
+fun ErrorTextFieldItemPreview() {
+    ShopAppTheme {
+        ErrorTextFieldItem(
+            errorMessage = "Error message",
+            testTag = "tag"
+        )
+    }
+}
