+++ app/src/main/java/com/example/shopapp/composable/HomeScreen.kt
@@ -1,0 +1,62 @@
+package com.example.shopapp.composable
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Scaffold
+import androidx.compose.material.rememberScaffoldState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.shopapp.ui.theme.ShopAppTheme
+
+@Composable
+fun HomeScreen() {
+    val offers = listOf(
+        "All clothes for women now 10% cheaper",
+        "All clothes for men now 15% cheaper",
+        "All shirts 20% cheaper with code SHIRT20",
+        "Buy two pairs of pants for the price of one",
+        "13% off for purchase above 200$"
+    )
+    val scaffoldState = rememberScaffoldState()
+
+    Scaffold(
+        topBar = { HomeTopBar() },
+        bottomBar = { BottomBar() },
+        scaffoldState = scaffoldState,
+        modifier = Modifier
+            .fillMaxSize()
+            .background(MaterialTheme.colors.background)
+            .padding(horizontal = 10.dp)
+    ) { innerPadding ->
+        Column(
+            modifier = Modifier
+                .padding(innerPadding)
+        ) {
+            LazyColumn(
+                modifier = Modifier
+                    .fillMaxSize(),
+            ) {
+                itemsIndexed(offers) { _, offer ->
+                    OfferItem(
+                        text = offer
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Preview
+@Composable
+fun HomeScreenPreview() {
+    ShopAppTheme {
+        HomeScreen()
+    }
+}
