+++ app/src/main/java/com/example/shopapp/composable/HomeTopBar.kt
@@ -1,0 +1,53 @@
+package com.example.shopapp.composable
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.ShoppingCart
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.shopapp.ui.theme.ShopAppTheme
+
+@Composable
+fun HomeTopBar() {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(MaterialTheme.colors.background)
+            .padding(vertical = 10.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Text(
+            text = "Shop Name",
+            fontWeight = FontWeight.SemiBold
+        )
+
+        Icon(
+            imageVector = Icons.Outlined.ShoppingCart,
+            contentDescription = "Cart",
+            modifier = Modifier
+                .clickable {}
+        )
+
+    }
+}
+
+@Preview
+@Composable
+fun HomeTopBarPreview() {
+    ShopAppTheme {
+        HomeTopBar()
+    }
+}
