+++ app/src/main/java/com/example/shopapp/composable/LoginScreen.kt
@@ -1,0 +1,135 @@
+package com.example.shopapp.composable
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.shopapp.ui.theme.ShopAppTheme
+
+@Composable
+fun LoginScreen() {
+
+    val email = "email@email.com"
+    val emailError = ""
+    val password = "password"
+    val passwordError = ""
+    val isLoading = false
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(MaterialTheme.colors.background)
+            .padding(horizontal = 20.dp),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.SpaceEvenly
+    ) {
+        Text(
+            text = "Hello",
+            color = MaterialTheme.colors.secondary,
+            fontSize = 32.sp,
+            fontWeight = FontWeight.SemiBold
+        )
+
+        Column {
+            ShopTextFieldItem(
+                text = email,
+                label = "email",
+                placeholder = "placeholder",
+                testTag = "email",
+                isError = emailError != null,
+                onValueChange = {},
+                keyboardOptions = KeyboardOptions(
+                    keyboardType = KeyboardType.Email
+                )
+            )
+
+            if(emailError != null) {
+                ErrorTextFieldItem(
+                    errorMessage = emailError,
+                    testTag = "errorTag"
+                )
+            }
+
+            Spacer(modifier = Modifier.height(10.dp))
+
+            ShopTextFieldItem(
+                text = password,
+                label = "password",
+                placeholder = "placeholder",
+                testTag = "tag",
+                isError = passwordError != null,
+                onValueChange = {},
+                keyboardOptions = KeyboardOptions(
+                    keyboardType = KeyboardType.Password
+                ),
+                visualTransformation = PasswordVisualTransformation()
+            )
+
+            if(passwordError != null) {
+                ErrorTextFieldItem(
+                    errorMessage = passwordError,
+                    testTag = "tag"
+                )
+            }
+        }
+
+        ShopButtonItem(
+            text = "Login",
+            testTag = "tag",
+            onClick = {}
+        )
+
+        Row(
+            modifier = Modifier
+                .fillMaxWidth(),
+            horizontalArrangement = Arrangement.Center
+        ) {
+            Text(
+                text = "Don't have an account?"
+            )
+
+            Spacer(modifier = Modifier.width(5.dp))
+
+            Text(
+                text = "Sign up",
+                color = MaterialTheme.colors.primary,
+                fontWeight = FontWeight.SemiBold,
+                modifier = Modifier
+                    .clickable {}
+                    .testTag("tag")
+            )
+        }
+    }
+
+    if(isLoading) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .testTag("tag"),
+            contentAlignment = Alignment.Center
+        ) {
+            CircularProgressIndicator()
+        }
+    }
+}
+
+@Preview
+@Composable
+fun LoginScreenPreview() {
+    ShopAppTheme() {
+        LoginScreen()
+    }
+}
